Use Haexpress to encode Exprs
This simplifies the code quite a bit by separating concerns about Exprs in
another library.  Code can still be further simplified as I move relevant stuff
into Haexpress, but this is a start.

All tests are passing.  There is one small change in behaviour on the
"speculate-reason" example.

Unfortunately, runtime increased by roughly 2x.  I believe the culprit is the
"Instance" module, which now stores instance information in a flatter
structure, so Speculate takes longer traversing it.  I think this can be fixed
by "filtering" only what is needer earlier or using a Map, but I prefer trying
the former first.  I'll work on improving that on a later commit.