Use ListableExpr from Haexpress (failing tests)
Uses ListableExpr from Haexpress, this adds failing tests.

It seems when I moved the error treatment out of the Eq instance I have changed
behaviour a bit, causing Speculation to be slower specially in the insertsort
example.  The insertsort example may have a lot of error values due to the
application of head.