Revert "avoid normalize on consider, use isRootNormalE"
This reverts commit 81b1750938fd349414f29f23ecaa02e5d0d0ef9f.

Here's a stretchy example why the change was reverted.

Take the following expression:

10 * x

Depending on a number of factors, it may rewrite to:

x + x + x + x + x + x + x + x + x + x

The second is far bigger than the first, so we may end up not getting it
enumerated soon.  This could even negatively impact performance as we would
lose an equation until that bigger size.